Change filter component and fix accessibility.
