--- frontend/app/assets/styles/searchFilterPanel.css
@@ -1,21 +1,0 @@
-.search-filter-panel {
-	display: flex;
-	flex-direction: column;
-	gap: var(--space-m);
-	width: 100%;
-	padding: var(--space-m);
-	margin: 0;
-	border: 1px solid var(--borders-input);
-	border-radius: var(--radius-m);
-	background-color: var(--surface-card);
-}
-
-.search-filter-panel__field {
-	display: flex;
-	flex-direction: column;
-	gap: var(--space-xs);
-}
-
-.search-filter-panel__label {
-	color: var(--text-secondary);
-}
